Show projects in the selector by default
Project items should be shown unless there are explicit
"Visible To" entries that prevent it.
Project creators should also be able to see their projects
no matter what is inside Visible To.
If there's any issue with deserialization then we should also
fallback to the way it works now meaning the project will be
shown to the user.

Fixes: CMS-16369